Changed names of all test files
