+++ EZSwiftExtensionsTests/NSURLTests.swift
@@ -1,5 +1,5 @@
 //
-//  EZSwiftExtensionsTestsNSURL.swift
+//  NSURLTests.swift
 //  EZSwiftExtensions
 //
 //  Created by furuyan on 2016/03/02.
@@ -9,7 +9,7 @@
 import XCTest
 @testable import EZSwiftExtensions
 
-class EZSwiftExtensionsTestsNSURL: XCTestCase {
+class NSURLTests: XCTestCase {
     func testQueryParameters() {
         let url = NSURL(string: "http://ajax.googleapis.com/ajax/services/search/web?v=1.0&q=facebook")
         if let queryParameters = url?.queryParameters {
