+++ EZSwiftExtensionsTests/StringTests.swift
@@ -1,5 +1,5 @@
 //
-//  EZSwiftExtensionsTestsString.swift
+//  StringTests.swift
 //  EZSwiftExtensions
 //
 //  Created by Valentino Urbano on 29/01/16.
@@ -9,7 +9,7 @@
 import XCTest
 @testable import EZSwiftExtensions
 
-class EZSwiftExtensionsTestsString: XCTestCase {
+class StringTests: XCTestCase {
     var string: String!
 
     override func setUp() {
